+++ samplers/jaegerremote/internal/utils/rate_limiter.go
@@ -17,7 +17,6 @@
 package utils // import "go.opentelemetry.io/contrib/samplers/jaegerremote/internal/utils"
 
 import (
-	"math"
 	"sync"
 	"time"
 )
@@ -53,7 +52,7 @@ type RateLimiter struct {
 func NewRateLimiter(creditsPerSecond, maxBalance float64) *RateLimiter {
 	return &RateLimiter{
 		creditsPerSecond: creditsPerSecond,
-		balance:          math.Min(creditsPerSecond, maxBalance),
+		balance:          maxBalance,
 		maxBalance:       maxBalance,
 		lastTick:         time.Now(),
 		timeNow:          time.Now,